refactor(mining): dedupe day helpers into utils/day
user-context-builder had private copies of isSameDay/getDayBoundaries
identical to the ones in task-miner/helpers. Move both to
src/main/utils/day.ts so neither service imports from the other.
