+++ src/main/services/task-miner/helpers.ts
@@ -50,29 +50,6 @@ export function computeEpisodeWindow(activities: TimeBounded[]): {
   }
 }
 
-export function isSameDay(a: number, b: number): boolean {
-  const da = new Date(a)
-  const db = new Date(b)
-  return (
-    da.getFullYear() === db.getFullYear() &&
-    da.getMonth() === db.getMonth() &&
-    da.getDate() === db.getDate()
-  )
-}
-
-export function getDayBoundaries(daysBack: number): {
-  start: number
-  end: number
-  label: string
-} {
-  const now = new Date()
-  const day = new Date(now.getFullYear(), now.getMonth(), now.getDate() - daysBack)
-  const start = day.getTime()
-  const end = start + 24 * 60 * 60 * 1000 - 1
-  const label = `${day.getFullYear()}-${String(day.getMonth() + 1).padStart(2, '0')}-${String(day.getDate()).padStart(2, '0')}`
-  return { start, end, label }
-}
-
 export function serializeActivities(activities: ActivityDetail[]): object[] {
   return activities.map((a) => ({
     id: a.id,
